Improving stream verification
Our previous stream verification step was problematic in that if
the forall loops used to implement the computation and verification
loop skipped over elements, the verification would still pass.
Worked around this by using three forall loops: one to do the
computation, a second to compute the compuation for verification
purposes (this is "A-hat" in the official HPCC specification),
and then a third to do the infinity norm of the difference between
the two.  With this technique, if either one of the two computation
foralls has a problem, the verification is almost certain to fail.
If we wanted to be even safer, we could check that no elements of
A were 0.0 (since that's impossible or close to it given the
random numbers populating the vectors), but given that we're
meeting the reference version's requirements, I stopped here in
the interest of keeping the code from getting too big.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14964 3a8e244f-b0f2-452b-bcba-4c88e055c3ca