multi-task per locale versions of stream and ra
This changes the BlockDist to take an optional argument
indicating the number of tasks per locale to use, set to
the minimum number of cores per locale on the set of
locales to which the distribution is mapped.  Note that
this does not seem to be optimal for stream and ra 
(probably due to the polling thread in both cases and
incoming threads in RA's case), but it still makes sense
as a default ("1" being the other obvious choice).

Stream and RA modified to take advantage of this, and
each now has a config const that is used when setting
up the Block1D distributions that it uses.

I have seen some core dumps and problems at certain
numbers of threads, but none using 1 task per locale,
so believe that this is a step forward, if an unstable
one.  If others see the same problems I am (that is,
if it isn't due to other miscellaneous changes in my
directory), that would be a good reason to make the
default number of tasks be "1".



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14966 3a8e244f-b0f2-452b-bcba-4c88e055c3ca