Make GASNET_MAX_SEGSIZE only get propagated by pbs if set
It turns out that if GASNET_MAX_SEGSIZE is propagated by the
qsub command when it is not set, it causes problems for GASNet.
A zero value must be getting propagated in that case.  I thought
I had tested that previously, but must not have done so well.
With this change, it is only passed along if it is set in a
user's environment.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14969 3a8e244f-b0f2-452b-bcba-4c88e055c3ca