Renaming platform xt-cnl to xt-cle
"Cray Linux Environment" is marketing's preferred term for
the XT's operating system, so retired the "Compute Node
Linux" acronym in favor of it.  The main change that I've
made here aside from renaming is that xt-cnl used to only
be used as a CHPL_TARGET_PLATFORM, whereas I've permitted
xt-cle to be used as a CHPL_HOST_PLATFORM as well (because
the term CLE is meant to encompass both login and compute
nodes).  I've made it so that when it is used as a host
platform, the default compiler is gnu.

I've spoken with David about these changes so that he can
update his xt cron scripts accordingly.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14977 3a8e244f-b0f2-452b-bcba-4c88e055c3ca