Re-enabling --no-local compilations
This is a stab in the dark -- I didn't want to back out
more of Samuel's last checkin than I had to, and started
by looking at the changes to ChapelArray.chpl since that
is where the compiler was complaining for --no-local calls,
and seem to have found that commenting out the increment
of _count in _initialize() causes things to work again.
(Fingers crossed).



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14984 3a8e244f-b0f2-452b-bcba-4c88e055c3ca