Adding a fully fragmented version of stream
This version adds a second coforall within the loop that was
manually fragmented across locales to emulate an OpenMP-level
of parallelism within the official HPCC entry.  From a couple
of spot-checks on my desktop, the performance is close enough
to our normal stream.chpl to make this a valid comparison
against the HPCC MPI + OpenMP Stream runs.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14997 3a8e244f-b0f2-452b-bcba-4c88e055c3ca