Making GASNet 1.12.0 the default
My testing of GASNet 1.12.0 indicates that it's
no worse than the previous version on the runs
I've tried, so checking it in here for now to
get some runs against it tonight.  If there are
any surprises (here or in peoples' personal use),
we can switch back to 1.10.0 by reverting this
checkin.  In either case, we'll want to blow
one of the two versions out of the release, if
not the svn tree prior to the release.

You may need to manually remove your
third-party/gasnet/install directory in order
to get this working.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15003 3a8e244f-b0f2-452b-bcba-4c88e055c3ca