Enabled the --destroy-value-type-vars flag by default when --local is also enabled.  This flag is otherwise disabled by default, and it can be turned off with --no-destroy-value-type-vars.  However, it cannot be enabled if --no-local is specified, or if comm!=none.  This will be corrected soon after the upcoming release.
Also removed the line of code that disabled this flag if a begin statement was
encountered.  I forgot this line was there, and it is no longer necessary, as
proper analysis is done on each individual variable to ensure it is safe to
destroy even in the presence of begin statements.

Removed a few .compopts files now that --destroy-value-type-vars is on by
default.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15006 3a8e244f-b0f2-452b-bcba-4c88e055c3ca