Finished pass through SVNLOG, updating CHANGES
Still need to sort and filter appropriately, but this should serve
as a reasonable reference for making sure the spec is up-to-date.
Let me know if you think I missed anything important.  It's hard
to maintain concentration through all those SVNLOG entries (particularly
those that don't summarize well in the header line... :P



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15007 3a8e244f-b0f2-452b-bcba-4c88e055c3ca