Updating REGRESSIONS based on last night's runs
This looks quite clean and fixes up the big three problems from
yesterday:

- valgrind runs
- --no-local tests
- leakedMemory2/3 on --local compilations

Let's consider the code frozen unless the group broadly agrees to
future changes.  There are a few mysterious regressions that it
would be nice to clear up before the release if there was an
easy fix but otherwise I think we're in good shape.  I'll send out 
a summary in a separate mail.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15009 3a8e244f-b0f2-452b-bcba-4c88e055c3ca