Fixing up PBS launcher for NCCS machines
Our previous check for whether or not we were using the NCCS
qsub wrapper was weak -- it didn't support a --version flag,
so if we tried to call --version on qsub and didn't get anything
useful back, we assumed it was NCCS.  This morning, the qsub
script changed to pass the --version flag on to the underlying
torque version of qsub which is problematic because not only
did it break our weak test, it also makes their qsub look like
torque (even though it requires/accepts different flags).

To get around this issue for the release while also being
more resilient to future changes, I simply cloned the launch-pbs
directory and added support for an additional launcher option
pbs-nccs that always assumes qsub is the NCCS version.  Added
mention of this to the README.launcher and README.xt-cle files.

The original launch-pbs/chpl_cfg_launch.c is essentially
unchaged except that I have added some #ifdefs to allow
it to be compiled in a debug mode that uses "0" as the
taskid in generating unique dotfiles and doesn't blow
them away once the job has completed.  I find myself
making these kinds of changes locally every time I debug
this stuff and thought it was time to commmit them.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15014 3a8e244f-b0f2-452b-bcba-4c88e055c3ca