Changed launcher commands to quote argv[] arguments
When a user's command line arguments contained spaces,
those spaces were not being transferred to the main
binary properly.  Here, I quote the arguments in question
in order to preserve the spaces.  This should fix the
regression on GASNet (the only launcher that we test
on a nightly basis) for configComplex and also make our
users' lives better.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15018 3a8e244f-b0f2-452b-bcba-4c88e055c3ca