Changing for to forall
It turns out when I did my stream-fragfrag timing
spot-checks yesterday, I was compiling --local.
With --no-local some spurious communication was
inserted after local blocks were broken up due
to (unused) labels.  Switching the loop from a
for to a forall (as we do in stream-fragforall)
fixes this issue.  Thanks to David for catching
this -- I'd assumed that erring on the side of
a for would be the safe thing to do.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15020 3a8e244f-b0f2-452b-bcba-4c88e055c3ca