Took a pass over the CHANGES file: filtering, sorting, reorganizing
If you get a chance to look over this and think I've forgotten to mention
something important, please let me know.  We covered a lot of ground in
the past 7-1/2 months and I tried to emphasize things that would most
affect the user and that were bigger ticket items.  But I could've easily
missed some important work.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15039 3a8e244f-b0f2-452b-bcba-4c88e055c3ca