added new primitive INIT_REF to set references to NULL
This new primitive sets a reference to NULL.  Using the primitive MOVE
to move 'nil' into the reference has the semantics of writing nil to
the dereferenced value.  This primitive replaces how this used to be
accomplished via the type dtNilRef and the global gNilRef.  This new
mechanism has the advantage of simplicity.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15074 3a8e244f-b0f2-452b-bcba-4c88e055c3ca