added gTypeDefaultToken symbol and gTypeDefaultToken type
This symbol replaces the use of nil as a default value for a function
argument such that if the actual is not specified, the formal will be
initialized to the default value for its type.  We used to use nil as
a signal that this should be done, but I believe that has complicated
the implementation of nil in general.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15083 3a8e244f-b0f2-452b-bcba-4c88e055c3ca