Updating REGRESSIONS based on... a backlog of runs
Our cron jobs had gone astray due to a combination of the system
upgrade and my .dotfiles disappearing, so this is the first time
that a lot of these tests have run in some time, resulting in
a bunch of regressions and new behaviors, the most egregious of
which should be fixed based on a few changes of Steve's today.
I'm too lazy to enumerate the others -- do a diff or refer to
email streams.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15086 3a8e244f-b0f2-452b-bcba-4c88e055c3ca