Fixing a bug in compiler.pl for xt-cle cross-compilation
In general, if CHPL_HOST_PLATFORM == CHPL_TARGET_PLATFORM
and CHPL_HOST_COMPILER is explicitly set, we infer
CHPL_TARGET_COMPILER to be CHPL_HOST_COMPILER if it isn't
explicitly set.  (For example, if compiling on linux64
for linux64, if you set the host compiler to intel, we
infer the target compiler to be intel as well).

Now that we use xt-cle to refer to the host and target
node environments of a Cray XT, this same logic was
firing, even though different compilers should typically
be used for the host vs. target on the XT.  For example,
gcc for the host and Cray's cc wrapping gcc for the
target.  This had the effect of causing a user who set
their HOST_COMPILER to pgi getting pgCC for the runtime
as well, rather than the cc wrapping PGI (or whatever
other compiler they had selected as their default).

For this reason, made the xt-cle case of compiler.pl
always infer the target compiler to be xt-cnl-$PE_ENV
if it isn't set by the user, and moved the logic for
inheriting the host compiler down into the extremely
generic "if this is some other platform" case of the
script.  This seems to be working correctly.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15089 3a8e244f-b0f2-452b-bcba-4c88e055c3ca