Updating REGRESSIONS file
No significant changes from last night, mostly refactoring and formatting.
It definitely seems to be increasingly likely that there's a race in
distribution-like code (distributions or our tests that resemble them) --
look for the tests that get 'attempt to dereference nil' in --no-local,
gasnet-everything, and gasnet-fast.  I guess the race could also be in
the tests themselves, but the stream failures are the most worrisome
ones...  Also, as mentioned yesterday, ra seems to be dumping core
consistently on GASNet for as-yet-undiagnosed reasons.


the case that there's some sort of race 



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15091 3a8e244f-b0f2-452b-bcba-4c88e055c3ca