Created a script to check for standard C header inclusion called grepstdchdrs, which compiler/Makefile invokes automatically.
This script can take one argument, which is passed on to the egrep command
instead of the -q switch that is normally given to it.  This allows one to,
for example, have it tell one what line numbers contain the offending #include
directives.

Also, if this script doesn't recognize the current working directory, it first
cd's to the $CHPL_HOME directory.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15108 3a8e244f-b0f2-452b-bcba-4c88e055c3ca