This fix allows compiled apps passed via PBS to have longer names.
This fix allows compiled apps passed via PBS to have longer names. The prior maximum was 8 which got formatted to Chapel-<app>. If the app name exceeded 8 characters, the program would be terminated. Now, the format is Chpl-<app>, and should the app name exceed 10 characters, it will be truncated.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15117 3a8e244f-b0f2-452b-bcba-4c88e055c3ca