Made -v flag on Chapel executable remove -q from aprun command
We use the -q flag on aprun in order to squelch an application
resources line that's printed at the end of a program's termination.
Unfortunately, it also seems to squelch important things like
'OOM killer' messages.  With this change, running with the -v
flag will remove the -q flag causing those messages to be
printed (as well as any others that -q was squelching).

Longer term, we've filed a bug report/feature request against
the fact that -q squelches the OOM killer message:

	id=746681 in bugzilla




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15122 3a8e244f-b0f2-452b-bcba-4c88e055c3ca