removed the creation of a redundant reference type (under --no-local)
Before this change, the compiler would create a new reference type for
every wide class it created.  This resulted in two unique yet
identical reference types because the reference type for the original
class would be changed into a reference to the wide class when all
classes are changed to wide classes.  This means that the reference
types of a class and a wide class were the same but different.

Now they are the same.  Instead of creating a new reference type, we
simply set Type::refType to point to the reference type for the class.
We also rename it to ref_wide_type instead of having one called
_ref_type and one called _ref__wide_type.

This cleans up the generated header code (chpl__header.h) nicely.  For
arrays/deitz/jacobi.chpl, the line count is reduced from 2091 to 1791
when compiling with --no-local.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15128 3a8e244f-b0f2-452b-bcba-4c88e055c3ca