renamed ref type for wide class to make it more consistent
Given a class type 'class', the following will often be created in a
multi-locale compilation:

classes
-------
  class      { ... }
  heap_class { wide_class value; }

wide class records
------------------
  wide_class      { int locale; class addr; }
  wide_heap_class { int locale; heap_class addr; }

references
----------
  ref_class -> wide_class*

wide reference records
----------------------
  wide_ref_class  { int locale; ref_class addr; }

Note that there is no reference to the class.  This is odd to me and
may make it difficult to narrow wide references.  I'm not sure if this
is the case for non-classes.

There are also neither references to heap classes nor references to
wide heap classes.  I'm unsure whether these are ever needed.  I'm
also not sure whether this is the case in the non-class case.  Classes
are different because, in addition to wide references to classes,
there are wide classes.

Given these three missing references, it is worth noting for
completeness that, if they existed, we would also potentially want
three more wide reference types, one for each of the reference types.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15129 3a8e244f-b0f2-452b-bcba-4c88e055c3ca