Add MPI communication layer implementation
This uses the Open MPI 1.3 pre-release (not included here). I configured
Open MPI with the flags:
  --prefix=/install/path --enable-mpi-threads --with-threads=posix
  --disable-mpi-f77 --enable-mpirun-prefix-by-default

Each locale has a polling thread that watches for communication requests from
other locales. Possible request types are get, put, fork, non-blocking fork
and terminate. Upon receiving a request, the polling thread does the requested
MPI send or recv, or launches a task to handle a fork.

Remote get and put are implemented with two MPI calls - a send to wake up the
polling thread and a recv or send of the actual data.

Blocking remote forks send a message to wake the polling thread, send an
argument to pass to the forked function, then receive a 0-length message to
indicate that the function is finished executing. Non-blocking forks are the
same except don't do the final receive.

After getting initialized, the "main" threads on all locales except 0 go
into a barrier, while locale 0 executes the chpl_main function. Once chpl_main
is finished, 0 also goes into the barrier, allowing shut-down to happen. To
shut down, each locale sends a message to its own polling thread, then tries
to lock a mutex that the polling thread owns. Once the polling thread is done
with all MPI calls, it unlocks the mutex to signal that it is now OK to call
MPI_Finalize.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15135 3a8e244f-b0f2-452b-bcba-4c88e055c3ca