Moved memory tracking code to its own file
This allows the different memory layers to share it.

Also, moved the locks used to serialize access to the memory tracking code
into the functions the locks protect. This way the memory layer doesn't need
to be burdened with handling tracking related locks.

Moved the initialization of these locks out of the threading init code and
into the memory tracking init code. Made chpl_mutex_init non-static to handle
this.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15139 3a8e244f-b0f2-452b-bcba-4c88e055c3ca