Updated REGRESSIONS file for the new year
Took a pretty thorough pass across the REGRESSIONS file as compared to cron
jobs that have been running in the new year.  In particular, I blew away a
bunch of spurious tests from 2008 since it has become less and less clear
which ones we've fixed and which ones we haven't.  Currently, only tests
that have been spurious in the past week (2009 + a few days) are listed.

Otherwise, did a general cleanup and resorting of the file, and for most
platforms, verified that our list of regressions matches what we're seeing
on a nightly basis (typically, I only look at the diff of new/resolved
errors).

Overall, my current biggest concerns in terms of predictability or
impact on this file are (in no particular order):

- occasional races in GASNet, one of which Steve was hunting down in
  the memory allocator code at the end of last year (deitz)

- other gasnet-everything regressions that have lingered (all)

- getting test_refine stabilized (bradc)

- getting ra-seq stable across platforms (bradc)

- getting memory-related tests cleaned up, primarily w.r.t. all the
  opts flags for the weekend runs (figueroa)




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15144 3a8e244f-b0f2-452b-bcba-4c88e055c3ca