Make test_local2 a future
This test expects an error when a remote int is read within a local block.
The optimization that passes by value scalars that are only read causes it
to become local, so the error doesn't happen.

Added another test that writes to the remote int. This causes the optimization
not to trigger, so the expected error happens in the local block.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15146 3a8e244f-b0f2-452b-bcba-4c88e055c3ca