Restored previous functionality to grep for standard C headers
This script can again take one argument, which is passed on to the egrep command
instead of the -s switch that is normally given to it.  If no argument is
specified, lines that have the word "include" followed by something other than
slash are removed from the output (though the -s switch already has this effect
on sunbeam).


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15150 3a8e244f-b0f2-452b-bcba-4c88e055c3ca