Add a check that the *_real binary exists in launch-amudprun
If the _real binary doesn't exist, issue an error. Add a test that moves the
launcher binary but not the _real one to check for this error.

This check could be moved out to be part of the launcher interface instead
of being specific to amudprun.

Modify the .preexec files for the localname tests to move both the launcher
and the _real files. This fixes those two GASNet regressions (hangs).


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15153 3a8e244f-b0f2-452b-bcba-4c88e055c3ca