Add a sanity check function to the launcher interface
It is called just before issuing the launch command. In most cases, it checks
that the _real binary exists and issues an error if not. For the two PBS
launchers, it also checks that the PBS script and the expect script exist.

Also, added launch-apsim, which is used as the default if the target platform
is x1-sim or x2-sim. This is for internal purposes only.

Changed a few Makefiles so that if "-sim" is in the CHPL_MAKE_TARGET_PLATFORM,
it is removed before using it for include paths.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15163 3a8e244f-b0f2-452b-bcba-4c88e055c3ca