Removing hand-baked C types
Since the compiler's inception, we've used hand-baked types
like "int32" and "int64" to store integers of various widths
-- presumably due to concerns at the time that stdint.h was
not sufficiently portable.  Yet the definitions of these
types have not been implemented portably.  We've had a test
that verifies that the compiler's types are the right size,
but it isn't run on platforms where we only test example
programs.

Removed this by switching over to the stdint.h types
throughout the compiler and the inttypes.h PRI/SCN macros
for printf/scanf control strings (which must be enabled
through a special #define in C++ code).  As a result of
this change, chpltypes.h now only defines our compiler's
complex64 and complex128 types and most #inclusions of
chpltypes.h can be and have been removed.

Removed the flag that tested integer sizes since that's
no longer necessary and the test that ran with this flag.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15173 3a8e244f-b0f2-452b-bcba-4c88e055c3ca