Adding literal format strings to printfs
To work around a warning that Jim Dinan is seeing with g++ 4.3.2,
converted (f)printfs of the form:

        printf(str);

to:

        printf("%s", str);

This isn't required by the C/C++ language definition, but isn't
particularly intrusive either.  If and when we start developing
against g++ 4.3.2, there's a flag that can be thrown to disable
these warnings if we find ourselves needing to create control
strings dynamically but run into these warnings.





git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15174 3a8e244f-b0f2-452b-bcba-4c88e055c3ca