Correct typos and small issues in the spec and add a few examples
In the list of binary operators that can result in parameters, removed the
unary not operators.

Changed "homogeneous" to "homogeneous tuple".

Removed a redundant rule in the grammar for block statements. Since statements
is optional, the empty block rule is unnecessary.

Added an example to the "Filtering Predicates in For Expressions" section.
Added a variable declaration example.

 


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15175 3a8e244f-b0f2-452b-bcba-4c88e055c3ca