Adding initial support for IBM xlc/xlC compilers
Setting up a Makefile.ibm file for use with IBM's xlc and xlC
compilers.  The depend rules don't work yet because there doesn't
seem to be a means to dump the dependences to stdout, which is
how our rules are currently written.  Need to do some more
refactoring around that.

Fixed a problem with an int vs. char return type that the IBM
compiler revealed.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15185 3a8e244f-b0f2-452b-bcba-4c88e055c3ca