Removing mta- and xmt-specific stdint.h header files
The XMT compiler group has implemented stdint.h in the more recent
versions of the compiler, which we're using for testing, and which
should be available to our users on or around the same time as our
next release.  Without this change, our stdint.h files conflict
with the built-in ones as a result of changes made during this past
week.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15190 3a8e244f-b0f2-452b-bcba-4c88e055c3ca