Greatly improved the leader/follower iterators for arrays distributed by the BlockDist distribution by re-privatizing the domain after executing the on-statement, and the array in the follower iterator. Without re-privatizing, we end up referring to this.dom/this, which refers to the domain/array on the locale that the leader iterator starts in, not on the locale within the on-statement.
This is an unsatisfactory solution because it uses a primitive.  This
code should be easier to write (and one should definitely be able to
avoid primitives) because it is the basis of user-defined
distributions.  There are several open questions.  Can we ignore
privatization in user-defined distributions outside of their setup?
How can we re-privatize more easily, either transparently or simply?
Does the compiler need to do something special?  One idea is to
automatically insert re-privatization code in all methods of a
privatized objects.

This looks like it will improve the scaling performance of the HPCC
STREAM benchmark.  So far, I'm seeing much better scaling even at only
64 locales with 3 tasks per locale (380.10 GB/s vs. 355.28 GB/s).
This change decreases the number of remote gets for each remote thread
on each iteration of the STREAM benchmark from 60 to 20 (about).  Many
of the remaining remote gets appear to be in the iterator class.

Note, I believe a similar improvement may be able to be made to the
domain leader iterator, but I have not yet investigated this.

Improved compiler handling of PRIM_GET_PRIV_CLASS.  Where the type
referred to in this primitive is widened, the compiler now only
attempts to widen it if it is not already wide.  This change was
necessary because this primitive is now used in more places in the
code.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15195 3a8e244f-b0f2-452b-bcba-4c88e055c3ca