Add some timings for the HPCC RA benchmark
These were built from a compiler last updated on 1/19/2009. The timings measure
the test test/studies/hpcc/2008/ra.chpl built with GASNet version 1.10.0 and
with GASNet 1.12.0.

With GASNet 1.10.0, there were significantly fewer crashes, and no timeouts.

In most cases, performance was slightly better using version 1.12.0, which is
the opposite of what was expected.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15198 3a8e244f-b0f2-452b-bcba-4c88e055c3ca