Added compiler support to re-privatize privatized objects that are stored in fields when they are accessed via PRIM_GET_MEMBER_VALUE. This has the effect of re-privatizing to the same degree that I was doing by inserting primitives and undesirable code in the modules. Cleaned up the module code as a result.  The net effect is that the distribution now looks almost identical to how it looked before.
In the compiler, added a new flag PRIVATIZED_CLASS that marks the
function _supportsPrivatization in the ChapelArray module.  This flag
is then added to all class types that support privatization if
privatization is enabled.  Then the compiler uses this to re-privatize
objects when they are accessed from fields.

The next step, as surmised by Brad, is to actually store the
privatized ID in the field and then re-privatize from that.  That
should eliminate one remote access per re-privatization.

Fixed a few bugs in privatization in which domains and distributions
were not fully replicated or re-replicated when changes occurred.
There are still issues, but this keeps all current tests working.
With this re-privatization work, we end up relying on privatization
more, making it more important that it is correct while also helping
with performance overall.

The main improvement to the modules is that we re-replicate (via a new
distributed domain/array function called reprivatize) when we assign
new values to a domain.

A second improvement is that the distribution field tasksPerLoc is
replicated in the replicated distributions.  However, if it is
changed, it currently is not re-replicated.

Also added a field pid to the distributed privatized array class.
This is similar to the pid field in the distributed domain class.  We
now rely on this field being set to contain the privatized object id
so that we can re-privatize.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15204 3a8e244f-b0f2-452b-bcba-4c88e055c3ca