Fixed use of C errno reportage
Our calls to the primitive get_errno were embedded in the middle of
a halt state, causing the C errno variable to get reset to 0 due to
successful printfs and such before making the call.  Hoisted these
calls into the top of the conditional in which the error condition
is caught in order to capture the string and then use it in the halt
statement.  For the case I ran into, this caused the errno message
to correctly change from "Success" to "Bad file handle".  This
should hopefully help with the other cases as well.





git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15205 3a8e244f-b0f2-452b-bcba-4c88e055c3ca