Improved support for comm diagnostics and debugging
Removed #ifdef disabling of comm diagnostics and debugging.  These are
now on by default if they are turned on in a Chapel program.  Before,
they had to be turned on in the runtime and the Chapel program.  I
don't believe that they will add much overhead if they are not used.

Removed old support for _startCommDiagnosis() and _stopCommDiagnosis().

This is now replaced by startVerboseComm() and stopVerboseComm().
When these are enabled, we'll print out a message whenever a remote
get, remote put, remote fork, or remote non-blocking fork occurs.

Also added functions startCommDiagnostics() and stopCommDiagnostics().
These are used to count remote gets, remote puts, remote forks, and
non-blocking remote forks.  The function getCommDiagnostics() returns
a local array declared over LocaleSpace with 4-tuple elements that
correspond to the number of remote gets, remote puts, remote forks,
and remote non-blocking forks that are encountered on each locale.

Added a simplified stream benchmark test case that uses comm
diagnostics to capture the current number of comm events to execute
stream on 2 locales with 1 task per locale to make sure that changes
to the block distribution do not make matters worse.

Note that comm diagnostics is not yet thread safe.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15207 3a8e244f-b0f2-452b-bcba-4c88e055c3ca