refocused and improved re-privatization optimization
Restricted re-privatization to fields in iterator classes because
regular classes and records should already refer to the privatized
object via its ID.  The iterator class runs into trouble because it is
created by the compiler and can thus directly access array and domain
classes, rather than array and domain wrapper records.

Changed the re-privatization optimization so that it works by
identifying privatized object fields in iterator classes, and then
replacing these fields by the integral private ID.  This removes extra
remote gets during re-privatization because the private ID no longer
has to be accessed from a privatized object class.

Reduced expected number of gets in streamCommCheck test and added
tasksPerLocale configuration constant for experimentation purposes.

The net effect of this check-in is to reduce by 3 the number of remote
gets (from 19 to 16) counted in the streamCommCheck benchmark.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15209 3a8e244f-b0f2-452b-bcba-4c88e055c3ca