added a test case of "remote value forwarding"
Added a test of using a value in an on clause.  This test makes sure
that a compiler optimization fires by using the comm diagnostics to
make sure there are 0 remote gets.

I'd like to rename the optReadOnlyRefArgs optimization to remote value
forwarding because this is what the optimization aims to do.  That it
currently keys off of reference arguments is indicative more of the
current implementation.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15210 3a8e244f-b0f2-452b-bcba-4c88e055c3ca