improved effectiveness of remote value forwarding
Added pragma/flag FLAG_DONT_DISABLE_REMOTE_VALUE_FORWARDING that can
be applied to functions that use sync variables in a structured way so
as not to interfere with the remote value forwarding optimization.
Added this pragma to several functions in the runtime.

Reduced expected number of gets in streamCommCheck test (16 to 14) and
improved test so that it prints out the diagnostics on a failure. This
can help debug non-determinism if there is any.

Added a second test of remote value forwarding that ensures the
compiler still remotely forwards a value into an on-statement even in
the presence of a cobegin statement nested in that on-statement.

Fixed a bug in remote value forwarding in which variables that are
written could be marked as read only and optimized incorrectly if they
are written to within a nested second on-statement, begin-block, etc.
This didn't show up before this change because there were so many
other reasons for this optimization to be disabled.

Made parallel/begin/deitz/test_begin_cobegin1 deterministic.  It
started to always fail with this checkin because remote value
forwarding is more robust and started to apply to the 'x' variable in
the second cobegin (before this change to the test).

Also modified multilocale/diten/needMultiLocales/nestedOn so that it
obeys our memory consistency model.  It started failing because of
remote value forwarding on s.  Since there were no synchronization
events in the first part of this cobegin block, remote value
forwarding applied.

Detailed Explanation and Discussion of String Writing
-----------------------------------------------------

Remote value forwarding optimizes the generated code so that values,
which are only read, can be communicated to a remote thread directly
(rather than through a wide reference which would require later
communication).  For example, in the code

  def main() {
    var x = foo();
    on RemoteLocale do begin {
      useValue(x);
    }
  }

without this optimization a wide reference to variable x would be
passed to the thread that implements the on-statement.  With this
optimization, the value of x is passed because x is never written to.
If x were written to, the wide reference would need to be passed so
that a remote put can handle the update.

This optimization is disabled in the presence of sync variables
because code like

  def main() {
    var x = foo();
    var s: sync int;
    on RemoteLocale do begin {
      s;
      useValue(x);
    }
    x = bar();
    s = 1;
  }

requires the update to x (assigned bar()) to be reflected in the use
of x on the remote locale.  Before this checkin, we disable whenever
there is any access to a sync or single variable within the
on-statement.

After this checkin, we don't disable this optimization if the sync or
single access is within a function with the 'dont disable remote value
forwarding' pragma.

For example, I added this pragma to the internal functions that apply
to _EndCount classes because these are used to implement coforall,
cobegin, and sync statements, but these should never disable remote
value forwarding.  That is, given code

  def main() {
    var x = foo();
    on RemoteLocale do begin {
      cobegin { ... }
      useValue(x);
    }
  }

there is no reason to disable the optimization if there is nothing
else within the code that implies synchronization.  I believe this is
consistent with our memory consistency model.

I also applies this pragma to the string write function which uses
sync variables to make sure that string writing is atomic.  I believe
this is fine because this is below the level of the user.  However, we
can have odd behaviour in the code

  def main() {
    var x = foo();
    var s: string
    on RemoteLocale do begin {
      s.write("1");
      useValue(x);
    }
    x = bar();
    s.write("2");
  }

since x could have the value returned by foo() when it is used in the
on-statement even if the string ends up as "21".

I specifically marked the string write function in order to enable the
optimization for the stream benchmark because a halt is used within
the on-statement and the halt function calls the string write function
(indirectly).  Another choice is to mark the halt function if we think
the synchronization of string writing should be part of the memory
consistency model.  The interesting question is whether this
synchronization should be allowed to fall below the radar, or whether
the user should be able to rely on it.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15214 3a8e244f-b0f2-452b-bcba-4c88e055c3ca