simplified remote value forwarding and fixed a bug
Removed one conditional check that used to be necessary to determine
that a reference argument was not safe to dereference but was never
sufficient, and became redundant on my last change (since the
optimization is disabled for all unhandled cases).

Fixed a bug in which an infinite recursion of isSafeToDeref could
occur in the case where two nested functions were mutually recursive.
Added a test case that failed before this fix.  The fix is to keep
track of the references that have already been determined to be safe
rather than (most of) the functions in which such references existed.

Added more comments.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15219 3a8e244f-b0f2-452b-bcba-4c88e055c3ca